+++ api/core/schemas/address.py
@@ -2,6 +2,7 @@ from pydantic import BaseModel, validator
 from typing import List, Dict, Optional
 
 from .geojson import GeoJson, Feature
+from .camadas import CamadaParam
 
 from config import DISTANCIA_PADRAO_MTS_GEOSAMPA
 
@@ -28,26 +29,6 @@ class AdressSearch(BaseModel):
 
     endereco: GeoJson
     camadas_geosampa: Dict[str, GeoJson]
-
-
-class CamadaParam(BaseModel):
-
-    alias: str
-    layer_name: str
-    distance: Optional[float]=DISTANCIA_PADRAO_MTS_GEOSAMPA
-
-    @validator('layer_name')
-    def validar_layer_name(cls, value)->str:
-
-        value = str(value)
-        if not value.startswith('geoportal:'):
-            value = f'geoportal:{value}'
-        
-        return value
-
-class CamadaParamInternal(CamadaParam):
-
-    geom_col: Optional[str]
 
 
 class AdressSearchParameters(BaseModel):
