+++ api/core/schemas/camadas.py
@@ -48,3 +48,23 @@ class DetalhesCamada(BaseModel):
         if value not in GEOM_TYPES:
             raise RuntimeError(f'Tipo de geometria inesperada: {str(value)}')
         return value
+
+
+class CamadaParam(BaseModel):
+
+    alias: str
+    layer_name: str
+    distance: Optional[float]=DISTANCIA_PADRAO_MTS_GEOSAMPA
+
+    @validator('layer_name')
+    def validar_layer_name(cls, value)->str:
+
+        value = str(value)
+        if not value.startswith('geoportal:'):
+            value = f'geoportal:{value}'
+        
+        return value
+
+class CamadaParamInternal(CamadaParam):
+
+    geom_col: Optional[str]
